Add sync before reboot to reduce lost data
This is motivated by a change in #16. The addition of the sync(2) call
in that PR was to ensure that a log file was written.

While the log file in that PR doesn't exist, calling sync(2) before
reboot(2) is good for reducing other data loss due to reboot(2) being
very ungraceful in restarting the system.