Be more precise on WDT pet timeout
The original code would pet the hardware watchdog (WDT) every 5 seconds
and on heart beat messages from Erlang. This worked for hardware
watchdog timeouts in the official Nerves systems.

This update breaks out the timeouts for the WDT so that they're separate
from the Erlang VM timer and not summarized in a select timeout. The
main point is to make it easier to communicate how timeouts work with
nerves_heart with other people.

A side effect of this change is that the WDT will be pet less frequently
than before on devices that have WDT pet times > 5 seconds (pet time is
less than the WDT timeout). For example, if the Erlang heart beat
interval is 60s and the WDT timeout is 120s, the WDT will be pet every
time Erlang sends a heart beat message. That will likely be >5s. For the
case where the WDT timeout is less than the Erlang heart beat interval,
nerves_heart will bet the WDT either 10 seconds before the WDT timeout
or half the WDT timeout, whichever is longer. This change is small
enough that it's doubtful if anyone would notice unless they instrument
the system and measure the delays.

This also updates the status printout to log the timeouts.