Make heart timer status more clear
This adds what Nerves heart thinks is left on the timers since not all
Linux kernel drivers report time left. It also prefixes anything from
the kernel watchdog with wdt to make it clear what status depends on the
kernel so that if it looks wrong, it's easier to point to the driver.

** The keys changed and are not backwards compatible **