Add minimum run time before even trying to reboot
This add support for `HEART_MIN_RUN_TIME`. If set to a number of
seconds, Nerves Heart won't let the system reboot for that amount of
time. For example, if you want the system to be up for 2 minutes so that
there's time to connect to the internet and get instructions, then set
`HEART_MIN_RUN_TIME` to 120. The max amount of time is 600 seconds or 10
minutes.