Add pmsg breadcrumbs to help debug watchdog reboots
This pulls out the logging code since it was getting long, and it's
basically a copy/paste from erlinit.

The logging code currently favors low frequency logging in that it will
retry opening log files rather than keeping handles around. It is
currently not intended to be a fast logger.